requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
astroid==2.13.2
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.1
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
colorama==0.4.6
constantly==15.1.0
cryptography==39.0.0
daphne==4.0.0
dill==0.3.6
Django==4.1.5
frozenlist==1.3.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
isort==5.11.4
lazy-object-proxy==1.9.0
mccabe==0.7.0
msgpack==1.0.4
multidict==6.0.4
openai==0.26.1
platformdirs==2.6.2
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pylint==2.15.10
pyOpenSSL==23.0.0
python-dotenv==0.21.1
redis==4.4.2
requests==2.28.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==23.1.1
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
wrapt==1.14.1
yarl==1.8.2
zope.interface==5.5.2
